feat: refactor checkout logic and namespace stripe under checkout
